refactor: move screens into packages
